+++ DeckStringDecoderKMPDemo/DeckStringDecoderKMPDemo/DeckStringDecoderKMPDemoApp.swift
@@ -1,0 +1,17 @@
+//
+//  DeckStringDecoderKMPDemoApp.swift
+//  DeckStringDecoderKMPDemo
+//
+//  Created by Sunset on 17/5/2026.
+//
+
+import SwiftUI
+
+@main
+struct DeckStringDecoderKMPDemoApp: App {
+    var body: some Scene {
+        WindowGroup {
+            ContentView()
+        }
+    }
+}